refactor(search): extract outbound fetch transport
